database/requirements.txt
Faker
psycopg2-binary
python-dotenv
